rename base.py to task.py to match the Task class
- Rename base.py to task.py in tasks directory
- Update all imports from .base to .task throughout codebase
- Update imports in tasks/__init__.py, cli/types.py, tests/utils.py
- Update imports in all task implementation files
- All tests passing (48 passed, 2 skipped, 14 deselected)
