+++ src/czbenchmarks/tasks/clustering.py
@@ -8,7 +8,7 @@ from czbenchmarks.types import ListLike
 from .types import CellRepresentation
 from ..metrics import metrics_registry
 from ..metrics.types import MetricResult, MetricType
-from .base import Task
+from .task import Task
 from .utils import cluster_embedding
 from .constants import N_ITERATIONS, FLAVOR, KEY_ADDED
 from ..constants import RANDOM_SEED
