+++ src/czbenchmarks/tasks/embedding.py
@@ -6,7 +6,7 @@ from ..tasks.types import CellRepresentation
 from ..types import ListLike
 from ..metrics import metrics_registry
 from ..metrics.types import MetricResult, MetricType
-from .base import Task
+from .task import Task
 
 logger = logging.getLogger(__name__)
