+++ src/czbenchmarks/tasks/single_cell/cross_species.py
@@ -3,7 +3,7 @@ from typing import List
 import numpy as np
 
 from ...constants import RANDOM_SEED
-from ..base import Task
+from ..task import Task
 from ...tasks.types import CellRepresentation
 from ...types import ListLike
 from ...metrics import metrics_registry
